Fix up helm and k3s
This PR changes the K3s server and Helm installs a bit, for helm we just
copy in the official helm install script and run it.

For K8s, rather than use cloud-init to start on boot, we can just
install and configure during our packer build, this way systemd will
take care of starting on boot.  We're not currently exposing an external
IP for the kube API so we can get away with this (we're expecting to run
bootstrap clusters locally on the VM).

We can change this in the future if need be, for now it serves our
purposes.